Move resover to separate class
